+++ tests/php/Unit/BladeRenderTest.php
@@ -1,20 +1,23 @@
 <?php
 
-declare(strict_types=1);
+declare(strict_types = 1);
 
-namespace AvtoDev\BackendToFrontendVariablesStack\Tests\Unit;
+namespace AvtoDev\Back2Front\Tests\Unit;
 
+use AvtoDev\Back2Front\Back2FrontInterface;
+use AvtoDev\Back2Front\Tests\AbstractTestCase;
 use Illuminate\Contracts\View\Factory as ViewFactory;
 use Illuminate\Contracts\Config\Repository as ConfigRepository;
-use AvtoDev\BackendToFrontendVariablesStack\Tests\AbstractTestCase;
-use AvtoDev\BackendToFrontendVariablesStack\Contracts\BackendToFrontendVariablesInterface;
 
+/**
+ * @covers \AvtoDev\Back2Front\ServiceProvider<extended>
+ */
 class BladeRenderTest extends AbstractTestCase
 {
     /**
      * {@inheritdoc}
      */
-    public function setUp()
+    public function setUp(): void
     {
         parent::setUp();
 
@@ -22,14 +25,12 @@ class BladeRenderTest extends AbstractTestCase
     }
 
     /**
-     * Rendering test.
-     *
      * @return void
      */
-    public function testRendering()
+    public function testRendering(): void
     {
-        /** @var BackendToFrontendVariablesInterface $service */
-        $service = $this->app->make(BackendToFrontendVariablesInterface::class);
+        /** @var Back2FrontInterface $service */
+        $service = $this->app->make(Back2FrontInterface::class);
         /** @var ViewFactory $view */
         $view = $this->app->make(ViewFactory::class);
         /** @var ConfigRepository $config */
@@ -58,14 +59,12 @@ class BladeRenderTest extends AbstractTestCase
     }
 
     /**
-     * Rendering test.
-     *
      * @return void
      */
-    public function testRenderCaching()
+    public function testRenderCaching(): void
     {
-        /** @var BackendToFrontendVariablesInterface $service */
-        $service = $this->app->make(BackendToFrontendVariablesInterface::class);
+        /** @var Back2FrontInterface $service */
+        $service = $this->app->make(Back2FrontInterface::class);
         /** @var ViewFactory $view */
         $view = $this->app->make(ViewFactory::class);